feat(averaged_collection): use f64 to prevent truncation

+++ averaged_collection/src/lib.rs
@@ -1,13 +1,13 @@
 pub struct AveragedCollection {
     data: Vec<i32>,
-    average: i32,
+    average: f64,
 }
 
 impl AveragedCollection {
     pub fn new() -> AveragedCollection {
         AveragedCollection {
             data: Vec::new(),
-            average: 0
+            average: 0.0
         }
     }
 
@@ -29,7 +29,7 @@ impl AveragedCollection {
         self.average = self.data.iter().sum();
     }
 
-    pub fn average(&self) -> i32 {
+    pub fn average(&self) -> f64 {
         self.average
     }
 }
